Rename wrappers to components
